backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
Django==3.1.7
django-cloudinary-storage==0.3.0
django-phone-field==1.8.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
Pillow==8.1.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.53.0
urllib3==1.26.3
django-cors-headers
